refactor: seperated Logics for report-tag endpoint
